requirements.txt
celery==5.2.7
django_filter==22.1
Django==3.2
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.1
gunicorn==20.1.0
psycopg2==2.9.4
python-dotenv==0.21.0
redis==4.3.4